Fix the flaky test of walreceiver
walreceiver is sensitive to any WAL write. After create a table
and insert some tuples, it doesn't run a vacuum. pg_class and
other catalog tables may be vacuumed in the background, which
causes some WAL traffic. To fix the flaky test, we run vacuum
again before the next test.

The return type of test_receive() doesn't match between the C
function and the declared function in sql file. We change the
return type for the SQL function.