[gpopt] Translate children of DXL Sequence in order.
We currently translate the last child of a sequence node before all
other children. This stands out as highly unusual because DXL node
translations normally have no order dependencies, and so they typically
are translated in order ("the normal for-loop"). There was a comment
rationalizing this, which only furthers confusion. We dug deep, and
found that there was no apparent reason we did that.

This patch removes the special handling of the last child, and rolls it
back into "the normal for-loop". This both results in clearer code, and
it removes a hurdle for the Postgres 12 merge where we needed a dynamic
table scan to be translated _after_ its corresponding static partition
selector. That is not great either, but we have a path forward to
eliminate the implicit order dependency there post-merge.

Co-authored-by: Alexandra Wang <walexandra@vmware.com>