Fix some postgres_fdw issues
* Compile and install postgres_fdw

* Enable postgres_fdw test
Modify postgres_fdw test cases to run it in GP.
1) Set up postgres_fdw progress by postgres_init.sh, postgres_fdw.sql
are run in gpdb, and it connects to postgres database by postgres_fdw.
We don't use lookback server as the remote postgres database.
2) Skip some EXPLAIN cases, as GP's plans sometimes are different with postgres
and the EXPLAIN cases are not essential.
3) "UPDATA" trigger can not work correctly in GP, just skip those related cases.

* INSERT RETURNING issue
For partition table, the tupleDesc between child partition
table and parent partition table may be different, because
parent partition table firstly drops a column, then adds a new
partition. The new child partition doesn't contain the dropped
column and shouldn't care about it when inserting data into the
new partiton. But when processing the RETURNING, we should use
the parent slot as there are something in planner to process the
dropped column.

For normal table, the values and nulls in parentslot (which are
inserted into the dest table) are not changed in ExecInsert
function, but for foreign table, there may be triggers on the
remote table, it can change the values, so we need update the
parentslot values for foreign table after inserting the data
into the remote dest table. Otherwise, the RETURNING expressions
will work on incorrect data.

* Fix postgresAnalyzeForeignTable