Use 64bit gxid for distributed transaction
Previously we use dtm_start_timestamp + 32bit gxid as gid. dtm_start_timestamp
is the starting timestamp of the coordinator. That normally works fine but it
has the issue of gxid overflow, which is not rare in real production
environment. It will be a more and more important issue when gpdb OLTP
performance is better and better. There was a patch trying to resolve the
issue.

	https://github.com/greenplum-db/gpdb/pull/10177

The patch handles wraparound, still keeping the 32bit gxid design. We use 64bit
gxid as gid instead in this patch as an alternative. For this idea, there were
some concerns during discussion on the github PR page.

- DistributedSnapshot is double sized, especially,
  DistributedSnapshot->inProgressXidArray might be large. For this, the
  size seems to be not a big burden on the coordinator node and if it
  that affects dispatching we could optimize using relative values during
  dispatching (e.g. use 'gxid - base_gxid').

- distributed log size: we do not need timestamp in it so the distributued
  entry size is same as before.

The rough design is as below,

We fetch a batch of gxid each time on the coordinator and when we used half of
them we wake up the dtx recovery to fetch one more batch and if there is no
available gxid, one of the backend would also try fetching a batch of gxid. The
fetch operation is wal logged and checkpoint will log the next gxid value. When
crash recovery happens on the coordinator we will jump into the next batch to
avoid gxid collision.

A new UDF gp_get_next_gxid() is provided to show the "next" gxid that could be
used.  Internally it is specified by the ShmemVariableCache->nextGxid variable.
This value is mostly not accurate after crash recovery or promote. When this
happens on the coordinator, as said we just jump to the next batch. If this
happens on the primary the value is usually out of date (read from previous
online checkpoint) but the value is not that useful so we won't update it until
the first mpp queries with gxid comes.

Some other notable changes are:

1. Use PgStartTime in gp_gettmid() which previously depends on the dtm starting
   timestamp. Using PgStartTime seems to be more correct per some offline
   discussion with people who work on related projects.  PgStartTime is
   dispatched in previous code.

2. Of course we do not need the dtm_start_timestamp.

3. gp_distributed_log and gp_distributed_xacts are modified since the gid
   column is not needed anymore.

There are still some left things which should or could be done as subsequent works.
- pg_resetwal
  How do we set a reasonable guessing value. Probably have to refer the standby
  node?  Else have to run gp_get_next_gxid() via utility mode on all segments
  and get the maximum value of them as the guessing value, but the UDF
  returning value is not accurate sometimes.
- dtx recovery: probably trigger the abort job only when needed
  (e.g. when dtx recovery knows there are prepared transactions need to abort
  related to retry abort failure or failover) or requested (via on/off guc or
  sql).
- distributed log: could we optimize using 32bit relative gxid value to save disk space?
- Maybe optimize the memory space for DistributedSnapshot->inProgressXidArray?
- TMGIDSIZE could be smaller when using hex characters.