Fix Orca CTAS for tables created with legacy hashops
For CTAS creating tables with non-legacy hashop distribution from tables
with legacy hashops, CTAS with Orca would distribute the data according
to the value of gp_use_legacy_hashops; however, it would set the table's
distribution policy hashop to the value of the original table. This
caused queries to give incorrect results as the distribution policy
mismatched the data distribution.

The data was being distributed correctly as we used the correct opfamily
internally in Orca when determining the distribution spec. However, when
populating the distribution policy value in the table in
CTranslatorDXLToPlStmt.cpp, we were trying to derive the opclass from
the opfamily. This doesn't work, and instead we need to pass the opclass
from the query through Orca and to the plan statement. Hence we use the
opfamily internally within Orca, and simply pass through the opclass to
populate the gp_distribution_policy value correctly.

This is aligned with the planner behavior and the table's actual
distribution.