gpstart: skip filespace checks for standby when unreachable
When the standby is unreachable and the user proceeds with startup,
gpstart fails to start when temporary or transaction files have been
moved to a non-default filespace.

To determine when the standby is unreachable fetch_tli was reworked to
raise a StandbyUnreachable exception. And the standby is not started if
it is unreachable.

Co-authored-by: Bhuvnesh Chaudhary <bchaudhary@vmware.com>