[WIP]Handle JSON ill-formed escapes
When deal with '\u' escape sequences in the JSON string, PG/GP throws an
error if the sequences cannot be de-escaped or represented as text type.
e.g.:
- Incomplete escape sequence like '\u0fe' (One digit is missing)
- Incorrect surrogate pairs
- '\u0000' which is a valid JSON encode but it cannot be represented
  with text type correctly.

PG tried to suppport '\u0000' case but failed and the '\u0000' was
disallowed by 451d280815 since the workaround introduced some
correctness issues which cannot be easily solved.

This commit tries to solve the issue by let user preserve the ill-formed
escape sequence or '\u0000' as it is in the JSON string field.
Two GUCs is introduced by this commit.

gp_json_preserve_ill_formed is a boolean value to enable the ill-formed
escape sequence handling. When it is true, the ill-formed sequence won't
cause a error while parsing JSON. Instead, the ill-formed sequence will
be used as the de-escaped text. For example:

postgres=# select json '{ "a":  "\u0000" }' ->> 'a' as it_is;
 it_is
--------
 \u0000
(1 row)

But in this way, the parsed JSON text cannot be distinguished from
source "\\u0000" (totally illegal one) and "\u0000" (the problematic one).
This is also one reason for commit 451d280815 .

gp_json_preserve_ill_formed_prefix is added to solve the above issue by
giving user control to decide what prefix to be added to the ill-formed
sequence string. e.g.:

postgres=# SET gp_json_preserve_ill_formed_prefix = "##";
postgres=# select json '{ "a":  "\u0000" }' ->> 'a' as it_is;
  it_is
----------
 ##\u0000
(1 row)

When gp_json_preserve_ill_formed is false, the behaviour is just like
before, an error will be reported when a escape sequence cannot be
handled.