Test gpfdists external table with hostname
On centos7, the system libcurl uses NSS instead of OpenSSL as backend
for TLS/SSL connections. Previously it will fail if hostname is used
in external table location, probably due to initialization issues.
This is fixed by commit 89a1211c. Modify original test case to use
hostname.

curl is used for connection testing in gpfdist_ssl_start. But it will
always fail since we are now using SSL. Change the condition to not
equal to 7. This code can be found at curl documentation:
https://curl.haxx.se/libcurl/c/libcurl-errors.html