[Travis CI] Bump cache size to 4G
Historically we've lowered the cache size to 500M (from the default of
5GB) because we thought that saves us time fetching and storing the
cache. But with GCC 8 on Linux the total object file size is a little
south of 3G. That means we lose more time in compilation cache misses
than the transfer time on the cache itself. This change alone speeds up
the GCC 8 on Ubuntu bionic job by more than 50%: from 32 minutes to 14
minutes.