Use double-parens params in pipeline definition.
Having double curly braces in a YAML file actually invalidates the
YAML-ness (yamility? Anyway, an open curly brace is the token for
beginning an "inline mapping"). I noticed this while trying to come up
with a regression test suite in a bigger refactoring to the pipeline
templates (gotta parse that YAML one way or another).

The double-parens syntax was introduced more than 3 years ago [1] to be
less rough around the edges.

[1] https://github.com/concourse/concourse/releases/tag/v3.2.0