Introduce a way to validate cluster state during regression tests
It is often observed in CI that a test that leaves the cluster in an
inconsistent state (e.g. primary-mirror pair is not in sync, or primary
has not finished crash recovery) cause several following tests in the
schedule to fail.  What's worse, the culprit test itself may be reported
as passed because its validation criteria did not include the state of
the cluster.  This has found to mislead debugging efforts, ultimately
waste of time.

To make debugging CI failures more efficient, this patch adds
--validate-system=TEST option to pg_regress.  It specifies the name of
the test that pg_regress would run to validate cluster state before
running an actual test from a schedule.  If the validation test fails,
further testing is aborted.  The validation is performed before running
a group of tests specified on one line in the schedule file.
Alternatively, validation could be performed before every single test,
but let's not get there until absolutely necessary.

When the cluster validation is found to fail, the culprit is in the
previously run test group.

The new option is turned off by default, leaving developer workflow
unchanged.  It is enabled only in CI jobs.

This patch is built on the ground work and analysis laid out in PR #9865
by Wu Hao.  In particular, Wu Hao found that the validation overhead is
small enough to make the proposed patch viable.