Add check-creds to set-pipeline command hint
This ensures that double-parens are either provided with a static value,
or have runtime substitution available at the credential store. While
we're at it, also use the long form of flags. This reads better, and is
IMO a lot more accessible to people who don't have Concourse on their
fingertips as often as we do.