Fix assertion failure with sorted plans with grouping sets.
For two-stage GROUPING SETS plans, we use an internal GROUPINGSET_ID()
expression to distinguish the rolled up rows. The input to the Finalize
GroupAggregate node is sorted using GROUPINGSET_ID(), but it is not in
the target list. It was nevertheless included in its pathkeys, which led
to confusion, if there was e.g. a Motion on top of the Finalize
GroupAggregate that tried to preserve the sort order.

Without assertions, I believe this generated the error:

ERROR: GroupingSetId found in non-Agg plan node (execExpr.c:872)

as we saw on the perf pipeline, although I didn't test that. With assertions
you got:

FATAL:  Unexpected internal error (tlist.c:393)
DETAIL:  FailedAssertion("!(list_length(dest_tlist) == list_length(src_tlist))", File: "tlist.c", Line: 393)