Fix register workfile in a freed workfile_set bug.
We used to free the workfile_set when all it's files are closed.
But in some cases, we use the workfile_set for the whole node execution
process. So, we should pin the workfile_set to avoid unexcepted free
operation. And manually remove it once we are sure we don't want use it
anymore.

This fix the issue: https://github.com/greenplum-db/gpdb/issues/10855.