Use the same code for generating two-stage DISTINCT plans as GROUP BY
Previously, we had a parse analysis step to convert DISTINCT into a
corresponding GROUP BY query, because the planner was smarter about GROUP
BY than DISTINCT. The DISTINCT planning code was still needed for more
complicated cases, though, where you had both DISTINCT and GROUP BY or
aggregates in the same query. The DISTINCT planning code was badly
dilapidated, there was broken commented-out remnants of "pre-unique"
optimization in the DISTINCT planner code, for example, that got broken
in the 8.4 and 9.6 merges, but we hadn't bothered to fix.

Clean that up. Instead of transforming DISTINCT into GROUP BY in the
parse analysis stage, re-use the GROUP BY planning code for DISTINCT
queries in the planner.

This removes the gp_eager_preunique GUC, that was previously used to
force two-stage DISTINCT plans. It was undocumented, only had an effect
in the limited cases that we didn't transform the DISTINCT to GROUP BY,
so we probably don't want it in that form. It would make some sense to
have a GUC to force multi-stage DISTINCT and aggregation, perhaps by
adding a penalty cost to one-stage plans, but that should be a separate
PR.

Co-authored-by: Jinbao Chen <cjinbao@vmware.com>