Skip Segments which are down for updating filespaces files on the host
If any segment host was down and the temporary or transaction filespace were
moved to a different filespace (i.e other than default), it used to do a
SSH to update the filespace content in a file on the down host and
will fail as the host is down. This will cause utilities by gprecoverseg
to fail even though if the segments on the down host are excluded from
the gprecoverseg input file.