Avoid repetitively constructing a QE's SegmentDatabaseDescriptor.whoami string.
1. Add a string field called defaultWhoami to SegmentDatabaseDescriptor. It stores
a string of QE identifier that contains backendPid but no slice index. whoami would
point to defaultWhoami whenever appropriate (e.g., when SegmentDatabaseDescriptor
is recycled or the QE has no slice index), so that cdbconn_setQEIdentifier() does
not have to re-construct this contents from scratch every time.

2. cdbcomponent_allocateIdleQE() should only call cdbconn_setQEIdentifier() if the
SegmentDatabaseDescriptor is newly created.

3. When constructing a new QE identifier string, cdbcomponent_allocateIdleQE() should
use a buffer whose initial size is 128 Bytes instead of 1024 Bytes, as QE identifier
are usually within that range.