Configure database with Python 3 by default
Configure needs to be run with the PYTHON variable set to the Python 3 binary
path to include PL/Python 3.

Ideally we would require the minimum version on configure to be Python 3, but
for now we're going to allow falling back to Python 2 as extensions and such
are not yet updated to Python 3.  We can tighten up that requirement at a later
date once there are no unupdated dependencies remaining.

For now it will make things a bit tricky if Python 3 is not installed, as the
configure will complete successfully and then tests will fail, instead of the
configure itself failing.

Co-authored-by: Ashwin Agrawal <aashwin@vmware.com>
Co-authored-by: Jamie McAtamney <jmcatamney@vmware.com>