Replace pickling with json and shlex in utilities
Pickling was previously used in several utilities when shelling out commands
and/or executing commands remotely, in order to avoid needing to escape strings
when passing them back to the master.  The actual string contents were largely
or wholly ASCII, so pickling was overkill for that purpose.

The semantics of byte strings in Python 3 breaks the pickling logic, so we've
taken the opportunity to simplify that whole logic stack.  Code that formerly
pickled strings now uses shlex.quote() to escape strings where possible and
serializes strings with json where that is insufficient, removing any helper
functions that are no longer necessary.

Authored-by: Jamie McAtamney <jmcatamney@vmware.com>
Authored-by: Tyler Ramer <tramer@vmware.com>

Removed unused or unecessary helper functions from gppylib

Shell escape function was unused and python 3 shlex.quote() function
should be used anyway.

canStringBeParsedAsInt was a silly helper function, and also failed to
actually complete the cast as string.

Authored-by: Tyler Ramer <tramer@vmware.com>