Partially fix CTAS pg_resgroup_get_status()
This is the 6X version of
https://github.com/greenplum-db/gpdb/pull/10856.

The query `INSERT INTO .. SELECT .. FROM pg_resgroup_get_status()` will
execute the UDF on a QE started on the master node, it's not right to
check `if (Gp_role == GP_ROLE_DISPATCH)`.

It's not a complete fix since 6X could not have a catalog change, a more
complex query might start the QE on a segment which has no such stats.
Better than nothing.