Fix register workfile in a freed workfile_set bug.
We used to free the workfile_set when all its files are closed.
But in some cases, we use the workfile_set for the whole node execution
process. So, we should pin the workfile_set to avoid unexcepted free
operation. And manually remove it once we are sure we don't want to use
it anymore. Otherwise, the freed workfile_set may be allocated to others.
This fixes the issue: https://github.com/greenplum-db/gpdb/issues/10855.

Also, take care of the empty workfile_set, free it when applicable.
Distinguish by the workfile_set life cycle, the workfile_set could be class
into several classes.
1. workfile_set for a temporary workfile with interXact false(see buffile.c).
Once the file closed, the workfile_set will also get freed.
And if current transaction aborted, the workfile's resource owner is
responsible for closing it if the workfile is leaked, which also removes
the workfile_set.

2. workfile_set for a temporary workfile with interXact true(see buffile.c).
Once the file closed, the workfile_set will also get removed.
Since the workfile doesn't register to a resource owner, WorkFileLocalCtl
is responsible for removing its corresponding workfile_set in
AtProcExit_WorkFile(which gets called when process exit).

3. workfile_set is pinned and saved to prevent unexpected free operation.
When the workfile_set gets pinned, after all the file in the workfile_set
gets closed, we don't free the workfile_set, cause we may register new workfile
later. Call workfile_mgr_close_set to explicitly remove the workfile_set.
To deal with aborting, since the resource owner of the workfiles in the
workfile_set will close the workfiles, and this operation will not
remove workfile_set when no file exists in the workfile_set. We have to free
these kinds of workfile_sets when transaction ends in AtEOXact_WorkFile.
If the workfile_set contains interXact workfiles, we still rely on
WorkFileLocalCtl to remove workfile_sets in AtProcExit_WorkFile. But currently,
we don't have this kind of usage.

NOTE: if a workfile_set gets created but no workfile register into it(in case of
error, RegisterFileWithSet not called), it will never have a chance to be removed
automatically through file close. Then we have to check and remove these
workfile_sets when the transaction end to prevent workfile_set overflow
in AtEOXact_WorkFile.

Co-authored-by: Gang Xiong <gxiong@pivotal.io>