pg_upgrade: remove check for unsupported distribution columntypes
This check was introduced in commit 242783 as the default hashing logic
changed in 6X. This check was introduced under the assumption that the
abstime, reltime, and tinterval types were being removed from 6X.  They
were not removed, however, and so using them as distribution keys is
actually supported.  Since we can't remove them now as that would be a
catalog change, the check can be removed instead.

Co-authored-by: Jamie McAtamney <jmcatamney@vmware.com>