Relfrozenxid must be invalid for append-optimized tables
Append-optimized tables do not contain transaction information in
their tuples.  Therefore, pg_class.relfrozenxid must remain invalid.
This is being done correctly during table creation, however, when the
table was rewritten, the relfrozenxid was accidentally set.  Fix it
such that diff with upstream is minimised.  In particular, the function
"should_have_valid_relfozenxid" is removed.

The fixme comments that led me to this bug are also removed.