Generate Right Outer Hash Join alternatives in Orca plans
With the Postgres 9.1 merge, the executor gained support for right outer
hash joins (ROJ). This PR creates ROJ alternatives in addition to existing LOJ
alternatives in cases where we believe a ROJ would be promising.

Since ROJs carry more risk than LOJs, we are more conservative when
generating these plans to account for cardinality misestimation.
Specifically, since we're projecting all the elements of the hash side,
if we underestimate the hash side we may cause spilling, whereas an LOJ
doesn't have that risk.

There were quite a few plan changes with this, most of which had no
measurable performance difference. Performance testing showed no
regressions and a couple of major improvements due to additional
partition selectors being available. One of the nice benefits of ROJ is
that we can create a PS on hash side, and reduce the partitions scanned
on the non-hash side. Optimization time does increase slightly in some
cases, but it's ~10% for complex plans and is only present when we have
outer joins.