gpstart: testing of improve handling of down segment hosts
The tests in commit be5d11e2286 contained a typo that caused the changes
in the Scenario "gpstart starts even if the standby host is unreachable"
to not properly cleanup after itself.  Though the test feature still
passes, this leaves a bug to be found later when more tests are added.