Use gp_command_count for gpfdist session ID
gpfdist intended to use gpdb-session-id+transaction-id+scan-number to
mark a gpfdist session. The scan-number was a process-static var which
would be increased when creating the plan.
This approach has a problem with PREPARE/EXECUTE. When executing a
prepared statement, the gpdb may use the cached plan to execute. In this
case, gpfdist thinks the query is for the same gpfdist session since the
scan-number doesn`t change (no new plan is created). Thus, gpfdist won`t
reset the file offset and will give the wrong results. This is a 100%
reproducible issue on gpdb5 & gpdb6. For gpdb7, due to the cost
calculation change in 93abe741cd6 , the problem still existed but not
always happened.

- In the gpfdist, `X-GP-SN` was treated as session id (see comments:
  X-GP-SN - session ID), but it was set with scan-number in the gpdb.
  The real `X-GP-SESSION-ID` had never been used in the code. To
  maintain the compatibility, keep the gpfdist as it is. But remove the
  `X-GP-SESSION-ID` since it is useless. Instead of putting scan-number
  into `X-GP-SN`, put the real gpdb-session-id into that section.
- The `X-GP-CID` won`t use the debug purpose `QEDtxContextInfo.curcid`
  anymore, `gp_command_count` is used as a better alternative -- it
  changes for every query including the `EXECUTE` statement.
- The planner maintained static scan-number is not needed anymore, the
  relevant code is removed.
- Added some cases to gpfdist. Those cases may not fail on gpdb7 even
  without this commit due to the cost calculation change. But they would
  fail on gpdb5 & gpdb6.

The problem was found with fix de9b0e2 and 78024fb (#10971 & #10984). When
backport this commit, those changes can be removed as well.

Fix #10976