Add Orca index only scan cost formula
Planner costs index-only-scan relative to index-scan. In the case of
index-only-scan, `RelOptInfo->allvisfrac` can be used to reduce the cost
when the visimap contains all-visible blocks. Thus, if index-only-scan
is possible, it will be favored.

In this commit, Orca also costs index-only-scan relative to index-scan
and scales the cost based on percentage of all-visible blocks in the
visimap. This is done by storing `pg_class.relallvisible` stats in
CDXLRelStats which is accessible during costing. In DXL this is added to
RelationStatistics:
```
<dxl:RelationStatistics ... RelPages="XX" RelAllVisible="YY"/>
```

Also in this commit, we update gpsd and minirepro to collect stats for
relallvisible. In doing this we also udpated the tools to use python3
sytnax.

And finally, it also adds a new option "index_only_scan_tests" in
cal_bitmap_test to callibrate index only scan cost/execution.