Allow Orca to choose index only scan on non-distribution columns
In order to produce an index only scan plan, all columns referenced in
the query must be covered by an index. This check is made in the
`CXformIndexGet2IndexOnlyScan::Transform`.

Issue is that distribution columns were being marked as used even when
they were not directly referenced. This is because they are implicitly
used to add motions. However, this meant that Orca would never consider
an index only scan if the distribution column was not covered by the
index.