Exciting hack to work around an Apple Clang bug
DbgStr() is designed to be unused until evaluated in a debugger. The
"used" attribute usually does the trick to convince a mainstream
compiler (GCC and upstream LLVM Clang) to generate code for DbgStr().
There are two issues with the "used" attribute, however:

1. It slows down compilation. Every translation unit that includes the
declaration of a derived class will take a hit generating DbgPrint() and
DbgStr(). But we really only need one copy of them. This is a fairly minor
issue.

2. Not so minor, However, is that AppleClang insists to omit code
generation for unused functions in a class template (this is borderline
bug, and there isn't a flag to disable such a "feature").

The best remedy is for each derived type, explicitly instantiate the base
template in the translation unit. This forces code generation even in
Apple Clang, and it speeds up compilation because we only emit code for
each explicit instantiation once.

(cherry picked from commit 914445a5a62bff4d99bd11543cd853e568d7de2a)